requirements.txt
PyPDF2
faiss-cpu
sentence-transformers
tqdm
requests
yaspin
